requirements.txt
Django
djangorestframework
django-cors-headers
python-frontmatter
PyYAML
pydantic
rich
